# Filters and print user lists