--- a/core/email_api.php
+++ b/core/email_api.php
@@ -534,6 +534,8 @@ function email_send_confirm_hash_url( $p_user_id, $p_confirm_hash ) {
 function email_notify_new_account( $p_username, $p_email ) {
 	$t_threshold_min = config_get( 'notify_new_user_created_threshold_min' );
 	$t_threshold_users = project_get_all_user_rows( ALL_PROJECTS, $t_threshold_min );
+	$t_user_ids= array_keys( $t_threshold_users );
+	user_cache_array_rows( $t_user_ids );
 
 	foreach( $t_threshold_users as $t_user ) {
 		lang_push( user_pref_get_language( $t_user['id'] ) );