This is a rewrite of project_get_all_user_rows(). This function is the core functionality of higher level functions that creates user lists based on project assignation, access levels, etc. Mainly for printing user option list for selectable input fields (print_api, filter_api, ...)

The new functionality of this code allows querying on multiple projects, with a per-project access-level check. All the users are retrieved and sorted with one single database query.
Also, the printing of user list does not use temporary space and bypass cache in order to save memory usage.
This helps solving a bug of poor performance detected when user count is very large.

Also, as a direct consequence of allowing multi-project queries, some filter related bugs may be solved now. Example: include the actual filtered projects in the user list search fields, for reporter, monitoring user, etc. 

Regarding "print_user_option_list" improvements, here are some test that i have made to measure times and memory usage, to get an idea:

Test case, 20k users
(It was run in debug mode, and discarding the first run that usually takes longer)
Measuring memory usage with "memory_get_peak_usage(true)", and time with "microtime(true)" on begin and end of print_user_option_list()

For one single project (public)    
  * on branch master:
    * standard: 96468992 (96 MB) (5.34 s)
    * disabling user rows cache: 33554432 (33 MB) (2.9 s)
    
  * on modified branch:
    * sorted by username, realname: 5505024 (5 MB) (2.27 s)
    * sorted by lastname: 99876864 (99 MB) (4.99 s)
    * idem, but disabling cache: 30932992 (30 MB) (2.89 s)

For ALL_PROJECTS best case:
  * master: 115605504 (115 MB)   10.20 s
  * mod: 5505024 (5 MB) 2.11 s