@cproensa 2015-11-05
I have reviewed the call to "user_cache_array_rows()" that was made inside "project_get_all_user_rows()"
For most of the callers it's not needed, especially on the cases of building very large user lists.

Instead it has been moved to the caller when necessary (for some email functions).