@cproensa 2016-04-24
close, to refactor when standard solution for SQL long clauses is implemented